Update patch set 1
Patch Set 1: Code-Review-1

Is it normal that we have so much generated code for Ecore? In org.eclipse.eef.edit/icons/full/obj16 we have icons for all the non-abstract concepts of Ecore, same thing in the plugin.properties etc...

That seems odd, isn't there an issue with a missing reference to the Ecore genmodel?

Patch-set: 1
Label: Code-Review=-1