Support legacy properties tabbed contributions in the EEF properties view
A new extension point is defined to contribute tab descriptor providers. These
providers can contribute tab descriptors to the EEF tabbed properties views
framework.
This extension point is used to contribute the tabbed properties views defined
thanks to the legacy framework in order to be able to show the legacy properties
in the new EEF tabbed properties view framework.

Change-Id: I1f502b3ffb4a241abf0dd74a1a7367f88939ec6b
Signed-off-by: Mélanie Bats <melanie.bats@obeo.fr>