Update patch set 3
Patch Set 3: Code-Review-1

(12 comments)

I thought you would add the IEEFTabbedPropertySheetPageContributor in the API and not just the contributorId. With just the contributorId we can do what we want in 99% of the uses cases (for example with Sirius) but the way the API of the EEFTabbedPropertySheetPage has been modified for Sirius, someone could use as the contributor something which is not a WorkbenchPart and while is is a bit far fetched, this use case could not be supported by your modifications.

Patch-set: 3
Label: Code-Review=-1