[498748] Do not mark a text field as dirty when rendering the current model state
When setting up the text field's content to the current state of the
model, we must call Text.setText(), which triggers the ModifyListeners.
But in this context we don't want to mark the widget as dirty, otherwise
leaving the page would trigger the associated edition tool (hopefully it
would be a no-op if the tool is properly idempotent, but it's still bad
for performance).

Bug: 498748
Change-Id: I90fb10c754a53147ba6e09593863bddc4cffc416
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>