[496058] Fix the computation of the ID in case of EEFViewCategory
With the new version of the default rules, the groups are created
according to the categories defined in the genmodel. In this case Sirius
provides an EEFViewCategory object to EEF. This object is just a POJO
representing an eObject and a category. This object overrides the
hashcode method. Then if we use the System.identityHashCode method to
get the hashcode in the computation of the ID we would do not have the
expected result for EEFViewCategory objects.

Bug: 496058
Change-Id: I51cf399ebbe7a759036103386ec56b8a041d5b94
Signed-off-by: Mélanie Bats <melanie.bats@obeo.fr>